refactor(webapp): kebab-case column-sizing file, group meta flags
Address review: rename the module to column-sizing.ts (repo uses
kebab-case for multi-word files) and place canGrow next to canShrink
in ColumnMeta with a short comment on each explaining the difference.
