+++ packages/webapp/src/types/react-table.d.ts
@@ -2,10 +2,10 @@ import '@tanstack/react-table';
 
 declare module '@tanstack/react-table' {
     interface ColumnMeta {
+        // Logs table: column grows to fill leftover row space, and is the last to shrink/truncate.
         canGrow?: boolean;
-        widthPercentage?: number;
-        // Logs table: column may shrink (and truncate) when the row runs out of horizontal space.
-        // Non-shrink columns keep their content width, so a grow column is the last to give up space.
+        // Logs table: column is allowed to shrink (and truncate) first when the row runs out of horizontal space.
         canShrink?: boolean;
+        widthPercentage?: number;
     }
 }
